requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
openai==1.51.0
aiosqlite==0.20.0
pydantic==2.9.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
apscheduler==3.10.4
